+++ RayTracerChallenge/Helper.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <cmath>
-const float EPSILON = 0.00001;
+const double EPSILON = 0.00001;
 
-inline bool epsilonEqual(float a, float b) { return fabs(a - b) < EPSILON; };
+inline bool epsilonEqual(double a, double b) { return fabs(a - b) < EPSILON; };
