+++ RayTracerChallenge/Ray.h
@@ -10,7 +10,7 @@ public:
 
 	Ray(const Tuple& _origin, const Tuple& _direction) : origin(_origin), direction(_direction){}
 
-	Tuple position(float t)const;
+	Tuple position(double t)const;
 	Ray transform(const Matrix& m)const;
