+++ RayTracerChallenge/Transformations.h
@@ -2,12 +2,12 @@
 #include "Matrix.h"
 #include <math.h>
 
-Matrix translate(float x, float y, float z);
-Matrix scale(float x, float y, float z);
+Matrix translate(double x, double y, double z);
+Matrix scale(double x, double y, double z);
 // angle is in radians
-Matrix rotationX(float angle);
-Matrix rotationY(float angle);
-Matrix rotationZ(float angle);
-Matrix shearing(float xy, float xz, float yx, float yz, float zx, float zy);
+Matrix rotationX(double angle);
+Matrix rotationY(double angle);
+Matrix rotationZ(double angle);
+Matrix shearing(double xy, double xz, double yx, double yz, double zx, double zy);
 
 Matrix viewTransformation(const Tuple& from, const Tuple& to, const Tuple& up);
