+++ RayTracerChallenge/Tuple.h
@@ -3,29 +3,29 @@
 
 struct Tuple {
 
-	Tuple(float x, float y, float z, float w);
+	Tuple(double x, double y, double z, double w);
 
-	static Tuple point(float x, float y, float z) {
+	static Tuple point(double x, double y, double z) {
 		return Tuple(x, y, z, 1.f);
 	}
 
-	static Tuple vector(float x, float y, float z) {
+	static Tuple vector(double x, double y, double z) {
 		return Tuple(x, y, z, 0.f);
 	}
 
 	bool isPoint();
 	bool isVector();
-	float magnitude()const;
+	double magnitude()const;
 	Tuple normalize()const;
-	float dotProduct(const Tuple& v)const;
+	double dotProduct(const Tuple& v)const;
 	Tuple crossProduct(const Tuple& v)const;
 
-	float x, y, z, w;
+	double x, y, z, w;
 };
 
 bool operator==(const Tuple& lhs, const Tuple& rhs);
 Tuple operator+(const Tuple& lhs, const Tuple& rhs);
 Tuple operator-(const Tuple& lhs, const Tuple& rhs);
 Tuple operator-(const Tuple& negate);
-Tuple operator*(const Tuple& negate, float scalar);
-Tuple operator/(const Tuple& negate, float scalar);
+Tuple operator*(const Tuple& negate, double scalar);
+Tuple operator/(const Tuple& negate, double scalar);
